+++ web/src/views/Profile/index.js
@@ -27,10 +27,6 @@ import { useSelector } from 'react-redux';
 import EmailModal from './component/EmailModal';
 import Turnstile from 'react-turnstile';
 
-import { Box, Typography, ClickAwayListener, Tooltip } from '@mui/material';
-import Chip from '@mui/material/Chip';
-import axios from 'axios';
-
 const validationSchema = Yup.object().shape({
   username: Yup.string().required('用户名 不能为空').min(3, '用户名 不能小于 3 个字符'),
   display_name: Yup.string(),
@@ -125,71 +121,7 @@ export default function Profile() {
     }
     loadUser().then();
   }, [status]);
-const [modelsByOwner, setModelsByOwner] = useState({});
-  const fetchModels = () => {
-    axios.get('/api/token/?p=0')
-      .then(response => {
-        const keys = response.data.data;
-        const validKey = keys.find(item => item.status === 1);
 
-        if (!validKey) {
-          throw new Error('请启用一个Key,用来查询可用模型');
-        }
-
-        const key = 'sk-' + validKey.key;
-        return axios.get('/v1/models', { headers: { Authorization: `Bearer ${key}` } });
-      })
-      .then(response => {
-        const models = response.data.data;
-        const modelsByOwner = models.reduce((result, model) => {
-          const owner = model.owned_by || '未分类';
-
-          if (!result[owner]) {
-            result[owner] = [];
-          }
-
-          result[owner].push(model.id);
-          return result;
-        }, {});
-
-        setModelsByOwner(modelsByOwner);
-      })
-      .catch(error => {
-        // 如果出现错误，就显示提示
-        setModelsByOwner({ 错误: [error.message] });
-      });
-  };
-
-
-
-  useEffect(() => {
-    fetchModels();
-  }, []);  // 添加空数组作为依赖，以确保这段代码只在组件载入时执行一次
-
-  // 模型复制操作
-  const ModelChip = ({ model }) => {
-    const [open, setOpen] = useState(false);
-
-    const handleClick = () => {
-      navigator.clipboard.writeText(model);
-      setOpen(true);
-    };
-
-    const handleClose = (event) => {
-      if (event) {
-        event.stopPropagation();
-      }
-      setOpen(false);
-    };
-
-    return (
-      <ClickAwayListener onClickAway={handleClose}>
-        <Tooltip open={open} onClose={handleClose} title="已复制" placement="top">
-          <Chip label={`${model}`} onClick={handleClick} sx={{ '&:hover': { transform: 'scale(1.1)', transition: 'transform 0.3s' } }} />
-        </Tooltip>
-      </ClickAwayListener>
-    );
-  }
   return (
     <>
       <UserCard>
@@ -257,18 +189,6 @@ const [modelsByOwner, setModelsByOwner] = useState({});
                 </Grid>
               </Grid>
             </SubCard>
-            <SubCard title="可用模型">
-              {Object.entries(modelsByOwner).map(([owner, models]) => (
-                <Box key={owner}>
-                  <Typography variant="h6" sx={{ marginBottom: '10px' }}>分类：{owner}</Typography>
-                  <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: '10px' }}>
-                    {models.map(model => (
-                      <ModelChip key={model} model={model} />
-                    ))}
-                  </Box>
-                </Box>
-              ))}
-            </SubCard>
             <SubCard title="账号绑定">
               {turnstileEnabled ? (
                 <Turnstile
@@ -333,6 +253,7 @@ const [modelsByOwner, setModelsByOwner] = useState({});
                     onClick={() => {
                       setShowAccountDeleteModal(true);
                     }}
+                    display="none"
                   >
                     删除帐号
                   </Button>
